Renamed govuk_utils and moved email.TemplateAttachmentEmail into its own module in plea
