+++ apps/feedback/forms.py
@@ -2,11 +2,11 @@ from django import forms
 from django.forms.widgets import RadioSelect
 from django.utils.translation import ugettext_lazy as _
 
-from apps.govuk_utils.fields import DSRadioFieldRenderer
-from apps.govuk_utils.forms import (YESNO_CHOICES,
-                                    to_bool,
-                                    BaseStageForm,
-                                    SplitStageForm)
+from apps.forms.fields import DSRadioFieldRenderer
+from apps.forms.forms import (YESNO_CHOICES,
+                              to_bool,
+                              BaseStageForm,
+                              SplitStageForm)
 
 from .fields import ERROR_MESSAGES
