+++ apps/plea/tasks.py
@@ -9,7 +9,7 @@ from django.core.mail import get_connection
 from django.conf import settings
 from django.utils import translation
 
-from apps.govuk_utils.email import TemplateAttachmentEmail
+from apps.plea.attachment import TemplateAttachmentEmail
 
 from manchester_traffic_offences.celery import app
 from apps.plea.models import Case, CourtEmailCount, Court
