+++ apps/plea/tests/test_email_template_output.py
@@ -8,7 +8,7 @@ from django.forms.formsets import formset_factory
 from django.test import TestCase
 from django.utils import translation
 
-from apps.govuk_utils.forms import RequiredFormSet
+from apps.forms.forms import RequiredFormSet
 
 from ..email import send_plea_email
 from ..models import Court
